Switched to a byte-by-byte parser with less state. (#1)
Specifically, we avoid carrying around all the variables indicating whether
we're inside a string or not, etc. We also avoid an initial copy over to a
buffer as we can decide on the type immediately. This simplifies the code.

It requires a more up-to-date version of byteme, as we're using the new
PerByte class (and its parallel equivalent) to scroll through the Reader.