Support pre-supplied fields and names in the Contents. (#5)
The ReadCsv class is now soft-deprecated in favor of read/read_file and
ReadOptions, which is a more natural way of passing options. Both read and
read_file have overloads that accept a pre-configured Contents.

In the pre-configured Contents, either of names/fields can be pre-supplied. If
names are pre-supplied, the parser will check that the same names are present
in the CSV. If fields are pre-supplied, the parser will fill them directly
rather than creating them. This enables validators to work easily.

Bumped the CMake project version.